+++ libft/ft_strlcat.c
@@ -3,23 +3,23 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcat.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jaemikim <imyourdata@soongsil.ac.kr>       +#+  +:+       +#+        */
+/*   By: jam_min_2 <jam_min_2@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/30 03:58:36 by jaemikim          #+#    #+#             */
-/*   Updated: 2023/11/04 17:10:37 by jaemikim         ###   ########.fr       */
+/*   Updated: 2023/11/05 02:06:44 by jam_min_2        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-int				ft_strlen(const char *s);
+size_t	ft_strlen(const char *s);
 
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 
 {
-	unsigned int	i;
-	unsigned int	dstlen;
-	unsigned int	srclen;
+	size_t	i;
+	size_t	dstlen;
+	size_t	srclen;
 
 	i = 0;
 	dstlen = ft_strlen(dst);
