+++ libft/ft_strnstr.c
@@ -3,20 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strnstr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jaemikim <imyourdata@soongsil.ac.kr>       +#+  +:+       +#+        */
+/*   By: jam_min_2 <jam_min_2@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/30 09:32:59 by jaemikim          #+#    #+#             */
-/*   Updated: 2023/11/04 17:17:40 by jaemikim         ###   ########.fr       */
+/*   Updated: 2023/11/05 02:13:13 by jam_min_2        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-int	ft_strlen(const char *s);
-int	ft_strncmp(const char *s1, const char *s2, size_t n);
+size_t	ft_strlen(const char *s);
+int		ft_strncmp(const char *s1, const char *s2, size_t n);
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
-
 {
 	if (*needle == '\0')
 		return ((char *) haystack);
@@ -24,7 +23,7 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 	{
 		if (*haystack == *needle)
 		{
-			if (len >= (unsigned int) ft_strlen(needle))
+			if (len >= ft_strlen(needle))
 			{
 				if (ft_strncmp(haystack, needle, ft_strlen(needle)) == 0)
 					return ((char *) haystack);
